Module openethereum::db::impls::migration

Enums

Error

Migration related erorrs.

Constants

BATCH_SIZE

Defines how many items are migrated to the new version of database at once.

BLOOMS_DB_VERSION

A version of database at which blooms-db was introduced

CURRENT_VERSION

Current version of database models.

DEFAULT_VERSION

Database is assumed to be at default version, when no version file is found.

TO_V11

The migration from v10 to v11. Adds a column for node info.

TO_V12

The migration from v11 to v12. Adds a column for light chain storage.

USE_MIGRATION_TOOL

Until this version please use upgrade tool.

VERSION_FILE_NAME

Version file name.

Functions

backup_database_path

Database backup

consolidated_database_migrations

Migrations on the consolidated database.

consolidated_database_path

Consolidated database path

current_version

Reads current database version from the file at given path. If the file does not exist returns DEFAULT_VERSION.

default_migration_settings

Default migration settings.

exists

migrate

Migrates the database.

migrate_database

Migrates database at given position with given migration rules.

update_version

Writes current database version to the file. Creates a new file if the version file does not exist yet.

version_file_path

Returns the version file path.